move gather code into sourmash_lib.search
